Check for images in /_images as well as /_sources/images

+++ scripts/sort_toc_nested.py
@@ -114,7 +114,10 @@ for path in Path('content').rglob('index.rst'):
       print ("Successfully created the directory %s " % dest_path)
 
     # copy images folder to destination folder
-    shutil.copytree(root_path + "images", dest_path + "/images")
+    if os.path.exists(root_path + "images"):
+      shutil.copytree(root_path + "images", dest_path + "/images")
+    else:
+      shutil.copytree(str(path.parents[1]) + "/_images", dest_path + "/images")
 
     # copy index over
     shutil.copyfile(root_path + "index.mdx", dest_path + "/index.mdx")
